Update patch set 1
Patch Set 1: Code-Review+1

It looks ok.  Since this is a Java project, you do not have to use the mvn-phase* method for building.  We can fix that later.  Now I am waiting for LF to approve the Jenkins job creation for your repo.  I will merge after that is in place.

Patch-set: 1