Update patch set 1
Patch Set 1:

Marcin, I think that this change is deliberately overwriting the dependency to include it with different version (which is dangerous, but can serve as a temporary workaround when tested).

As I understand, this dependency version has to be upgraded to be able to build ARM64 images. It raises my question - do we really need to build ARM64 images? :)

Patch-set: 1