Update patch set 5
Patch Set 5:

> Patch Set 4: Code-Review-1
> 
> Good start.  Couple comments and questions below
> 
> - Kafka config handling is not consistent with rest of PRH config handling which done through CbsContentParser class
> - Not sure if I missed in my review, where is the Configuration to toggle PRH msg reprocessing to be enabled/disabled? i.e how will this behavior controlled via helm configuration
> - Based on proposal discussion earlier, thought SDK lib will be enhanced to support the kafka interfaces? If this is something planned to be addressed separetely, pls create a jira for tracking this work.

We did not use CbsContentParser since we assumed the CBS is not / will not be used anymore. Can we make this change for next release, as we are already on M3.

The PRH msg reprocessing is an optional mode that is implemented in a different Spring Profile. This profile can be activated if the Helm deployment template contains an environment variable: autoCommitDisabled: true. 
The current working version  which is in the default Spring profile is not affected by our changes.

The SDK changes will be addressed separately. The JIRA will be created.

Patch-set: 5